requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tensorflow>=2.13.0
keras-facenet>=0.3.2
retinaface>=0.0.13
cryptography>=41.0.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
scipy>=1.11.0
djangorestframework>=3.14.0
whitenoise>=6.6.0
